Fix NO_CHECK with space-separated values
Do not quote $NO_CHECK, let the shell split it into separate arguments.
"${NO_CHECK[@]}" would work if that would be an array, but it is a plain
string. On the other hand, when shell split it, elementIn() will get it
as an array in $@, so the quoting there is correct.

Fixes QubesOS/qubes-issues#5563